requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
jinja2
psycopg2-binary
motor